mock-consensus: 👷 block data is implicitly empty, can be appended (#4004)
fixes #3936. based upon #4002. see #3588.

this changes the type of the block builder's `data`, allowing it to be
implicitly kept empty. this is useful when e.g. fast forwarding a number
of blocks _(see #4002)_. this spares us the need to call
`with_data(vec![])`, and additionally means that the block builder is
never in an uninitialized state _(making #4003 needless)._

* #3936
* #4002
* #3588