parent 3a9db32d7696f29afaf9cf7a9a65012037dbdd5e author plaidfinch <finch@plaidfinch.net> 1710279332 -0400 committer finch <finch@plaidfinch.net> 1710280707 -0400
Permit validator voting across an airgap (#3985)

This addresses the first changeset described in #3813, permitting
validator definitions and votes to be signed separately and those
signatures to be broadcast from a potentially-unrelated penumbra
account.

This is a breaking change to the CLI for all workflows that use the
`validator vote` subcommand, because it redefines that command to
`validator vote cast` in order to make room for `validator vote sign`.
It is purely a client-side change, and as such is not
consensus-breaking.

- [X] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label.

Delete unused import

Make formatting a bit nicer, set validator vote sign to offline

Allow initializing a separate governance subkey

Add a validator command to emit the governance key

Use the governance subkey correctly when present, in SoftKMS mode

Use the governance key, if configured, in the validator template

Fix mistake where the governance key was overzealously used

It should only be used to sign *votes*, not *definitions*!

Fix bug loading wrong config file path in dkg deal for governance keys