staking: use block height to enforce unbonding delay (#3923)
Close #3738, this PR makes the unbonding mechanism based on block delay
rather than epochs. In practice, this means that a user will wait X
blocks for a validator pool to unbond (rather than Y epochs).

To achieve this, this PR:
- defines `StakeParamater::unbonding_delay` measured in blocks
- parameterize unbonding tokens based on a `start_height`
- deprecate `start_epoch_index` fields in delegate/claim actions
- strip `epoch_index` from validator `RateData`s
- generalize the unbonding token denom format: 
`2.013718unbonding_start_at_1540_penumbravalid1` (vs. `_epoch_XXX`)

Important point about the mechanism: we bind tokens to the starting
height of the epoch that they belong to. This let us avoid binding
transactions to specific block heights.