refactor(staking): 👯 component flag reexports StateWriteExt
the changes in `crates/core/component/stake/src/lib.rs` are the
important bit here.

`StateReadExt` and `StateWriteExt` are complimentary traits, and thus
often used together. if we are going to reexport `StateReadExt` in the
top-level namespace, we should include both.

changes in other files adjust imports so that we import both from the
same path, instead of `crate::{StateReadExt, component::StateWriteExt}`,
to clarify their semantic relation.