dex: hotfix to avoid discarding EventPositionExecution
Closes #4007

The bug results from a hazardous API I created. The `apply` method does _not_
apply the events to the underlying state, it extracts them and returns them to
the caller, who can silently discard them.

A better fix is outlined in the comment:

```
// We have to manually extract events and push them down to the state to avoid losing them.
// TODO: in a commit not intended to be cherry-picked, we should fix this hazardous API:
// - rename `StateDelta::apply` to `StateDelta::apply_extracting_events`
// - add `StateDelta::apply_with_events` that pushes the events down.
// - go through all uses of `apply_extracting_events` and determine what behavior is correct
```

However, I didn't do this in this commit, because my hope is that we can have
something easily cherry-pickable into a point release, which can be deployed to
unblock the DEX work. After applying it to both `main` and a release branch, we
can then do the fix for the bug class.

(cherry picked from commit 318d3d18b0add9436d259c9dbc0a9dc5a9c47e8e)