staking: change `compute_unbonding_height` api (#4041)
This PR modifies the staking module logic responsible for determining
the unbonding schedule, in particular it:
- returns the current height if a pool is unbonded
- change the `compute_unbonding_height` API to return an `Option<u64>`
to allow callers to detect an unbonded pool

One tricky aspect of our unbonding mechanism is that we want to:
1. Bind unbonding notes to the economic effect of validator misbehavior
for some window of blocks
2. Leave it to the application to decide when a note is allowed to be
claimed

One thing to note is that this mechanism works relative to the _current_
bonding state of the pool. This is generally fine for the protocol
security because this leads the application to "overshoot" the penalty
window range since it cannot know the exact moment a note was allowed to
unbond (since bonding the state could have changed). This means that
when processing a claim, we detect if the pool is unbonded, and if so,
we set the "allowed unbonding height" to the current height.