feat(tests): ⚡ mock consensus delegates, and undelegates to a validator
see #3995.

this test now delegates to, and then undelegates from, a validator.

this does not yet include all of the assertions we would like to
exercise in this test, but it does show that a validator correctly
progresses through the active and defined states.

in further work, we will fill out the other assertions described
in #3995, particularly w.r.t. voting power and validator uptime.