feat(pd): support archives for migrate and join
Enables opt-in archive generation when performing:

  * pd export
  * pd migrate

The goal is to provide a standardized bottling-up of pd state,
specifically the rocksdb directory. In the context of upgrades,
the `pd migrate` functionality is what's important:
we want the archived dir to contain both rocksdb data and the modified
genesis file and updated validator state.

Accordingly, `pd testnet join` is modified to support an optional
archive URL. If set, the remote tar.gz archive will be downloaded
and extracted, clobbering the cometbft config files that were fetched.
A remote bootstrap node is still contacted, to learn about peers,
otherwise the newly created node wouldn't be able to talk to the network.