staking: handle param change and unbonding edge case (#4056)
This fix a bug in the claim logic that would cause the node to crash if
a parameter change had occurred, or if an undelegation was marked to
have taken place after the unbonding height specified in `Unbonding {
unbonds_at_height }`. The latter can happen if a validator gets ejected
from the consensus set for inactivity, and is equivalent to an
`Unbonded` state.