dex: internalize `PositionManager::put_position`
This implements an API suggestion originally made by @cronokirby to expose context-specific methods for updating positions.

We maintain a single `put_position` method on the `Inner` trait so that we're
able to have a single point where we edit any necessary indexes, but this is
now not exposed to callers. Instead they call a scope-specific method that
calls `put_position` internally.

Currently there's still room to improve the API safety but we'll want to
potentially change the signature of `put_position` to ensure that this remains
efficient. It would be ideal if it were possible to do all reads up front and
concurrently. Further investigation on this point would be desirable.

This commit has a security-critical change in that the logic for position
withdrawals' balance commitment changes, and it'd be good to get review on that
part in particular.