ci: add wasm check for component crates
Back in [0] we removed the wasm crate from the monorepo, so that the web
team could own it and iterate more quickly. We also purged associated
testing, however, so we missed some breakage around the release of
v0.70.0. This check attempts to restore some visibility into the
potential of breaking APIs relevant to web work.

[0] https://github.com/penumbra-zone/penumbra/issues/3819