dex: improve `position_manager::Inner::update_position` API
This commit changes `update_position` so that the caller is responsible for
fetching the previous state and then passing it in. While this adds misuse
potential into the API, the method is only used by the `PositionManager` and it
allows the high-level modification methods to do their own invariant checks
without blocking on extra reads.