tests(app): 🏓 add voting power assertions to staking tests (#4133)
see #3995.

this addresses points 10b and 11a in #3995. this adds additional
assertions to the `app_can_define_and_delegate_to_a_validator` test,
which ensure that the validator's voting power changes as expected.

#### checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > test changes only