cnidarium: implement deferred commits via `StagedWriteBatch` (#4122)
Refactor `commit` into `prepare_commit` and `commit_batch`. this
allows a `StagedWriteBatch` to be prepared via `prepare_commit` and
the actual writing the `StagedWriteBatch` to disk to occur in
`commit_batch`. this allows us to get the root hash of the changes
without requiring the data to be fully committed.

closes  #4095

---------

Co-authored-by: Erwan Or <erwanor@penumbralabs.xyz>