ci: use nightly rust for mdbook installation
Docs builds broke after merge of [0], not because its settings were
incompatible, but because its inclusion of a `rust-toolchain.toml` file
busted the CI cache, and caused the `mdbook` dep to be freshly
installed, rather than retrieved from cache. We don't version-lock
mdbook and friends, and the latest version evidently bumped MSRV to
1.74, which is newer than what we set for MSRV. We should probably bump
MSRV too, but I'll do that separately. This commit is to unbreak the
docs builds.

[0] https://github.com/penumbra-zone/penumbra/pull/4145