tests(app): 🔶 add genesis content setters
this is some common glue code in integration tests for the staking
component. in order to smooth the inspection of component logic that
runs at epoch boundaries, the defaults for `AppState`'s inner settings
for e.g. epoch duration are overriden to shorter values.

as is, this requres stating object literals that are somewhat verbose,
even with `..T::default()` elision. this commit adds some small `with_`
helpers to allow test cases to manipulate these settings succinctly.