proto: add extra metadata fields to `SwapView.Opaque`
Many of the "opaque" variants of our view messages don't have any extra data
other than the raw action. This turns out to be insufficient, because often
there's additional data that's practically necessary to understand the public
parts of the action.

For instance, in this case, the `SwapView`, we want to correctly and nicely
render the public parts of the swap, which include:

- the input amounts, which are transparent;
- the pro-rata output amounts, which are computable given public information.

However, this information can't be easily rendered just from the underlying
`Swap`, because:

- it doesn't contain the asset `Metadata` about the values, only the amount and asset IDs;
- it doesn't contain the `BatchSwapOutputData` with the (public) results of the batch swap;
- it doesn't have precomputed output values, so the rendering would have to understand and replicate the pro-rata computations.

Instead, we should make use of the perspective/view mechanism to augment the
`SwapView.Opaque` with that data.

To complete this PR, we'll need to:

- [ ] Regenerate protos and fix any compile errors in the Rust code.
- [ ] Ensure that the tx+perspective => view transformation uses data from the perspective to fill in the new fields (this should be similar to the existing code, but for opaque swaps rather than just visible ones).

The web frontend's code for displaying simulated "public views" of the user's
transactions can probably make use of this functionality after the first step
(since it constructs the simulated view by attenuating the "full view" rather
than by regenerating with a fresh perspective), but we should fill in the
perspective generation for completeness and to ensure that the data we're
adding actually makes sense and is possible to generate from the perspective.