penumbra: remove `Builder` factory (#4169)
## Describe your changes

We use named tasks in `pd` and `cnidarium` via `tokio::task::Builder`.
This is a great feature that improves `tokio-console` rendering and has
proved valuable in the past. However, since tracing integration in tokio
is still unstable, using this requires a special cargo configuration to
run rustc with a special `tokio_unstable` flag.

We have received a lot of direct feedback that this is 1/ surprising 2/
cumbersome, and since it seem like there are no immediate plans to
stabilize tracing in tokio, we are removing it for now.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > Internal only.