tests(app): 🏃 add fast forward to next epoch helper
this adds an extension method that can be used to fast forward to the
next epoch. this is a very common part of test logic that exercises
the staking component.

this allows test cases to invoke an extension method to jump ahead,
without needing to import the clock component and write a loop
inspecting the current epoch after creating and executing an empty
block.