tests(app): 🔑 use `validator_identity_keys` accessor
this adds a `ValidatorDataRead::validator_identity_keys()` trait method.

frequently, test logic needs to keep track of validators by their
identity keys. this led to a bit of common glue code that would get the
latest snapshot, the set of all validator definitions, assert that only
one validator is present, and then take the identity key of that single
validator.

now we can just get the keys, bind that single entry to a variable, or
return an error if something went wrong.