dex: track total number of positions per pair (#4167)
## Describe your changes

Working towards achieving #4077.

This PR adds a `PositionCounter` extension trait which implements a
`TradingPair` scoped position counter. I am not totally sold on a `u16`
for the counter type, yet, but it's useful for testing anyway.

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> This is consensus-breaking because it introduces a new implicit
validation rule: a hard-limit on the number of position opened for each
trading pair (via counter overflow). It could also be consensus breaking
if there was a bug in the DEX that allowed closed positions to be
updated, but my assumption is that it's not the case.