staking: force the ultimate definition upload to disable a validator (#4178)
## Describe your changes

This PR adds a **new consensus rule** which forces validators with
`sequence_number = u32::MAX` to be in the state `enabled = false`. This
reduces the incentive to steal validator identity signing keys, because
it means that it's impossible to "lock" a validator in the `enabled`
state; the operator can always permanently disable their validator,
cutting off profits for the attacker.

## Issue ticket number and link

This is an implementation of issue #4170; see its description for more
details about the risk this mitigates.

## Checklist before requesting a review

- [X] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> **This is a consensus-breaking change**: it rejects validator
definitions which would previously have been accepted by the chain.

Co-authored-by: finch <finch@starlingcyber.net>