tests(app): 🙅 app rejects invalid auth signatures (#4182)
fixes #4050. see also #3588.

this introduces a test case that demonstrates that transactions'
validator definition actions must have a valid authentication signature,
or the validator will not be added to the set of known validators.

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > this is a test case.