tests(app): 💎 polish mock consensus test infrastructure
see #3588. follows #4184 and #4181.

this takes a pass through the shared, Penumbra-specific test
infrastructure for mock consensus tests. notably, this decomposes
`init_chain.rs`, which has now become somewhat redundant with the
existence of other more involved tests of e.g. validator uptime
tracking.

this also cleans up some unused imports, guards against future
occurrences of that issue (_sharing code in `tests/` files is awkward_),
and decomposes the `common/mod.rs` file into some distinct standalone
components.

this also belatedly removes the `common::start_test_node()` helper. at
some point (_i was unable to find the link_) it was suggested that we
refrain from a shared setup helper like that. this branch removes that
helper, and updates its call-sites.

this branch is largely code motion, and is intended to be a last bit of
cleanup as we prepare for #3588 to wind down. :heart:

---------

Co-authored-by: Henry de Valence <hdevalence@penumbralabs.xyz>