tests: 🤝 hoist test subscriber into a standalone library (#4200)
in #4193, @zbuc is working on some changes to the dex component. in the
`penumbra-app` crate, we have a helpful utility to define a tracing
subscriber that correctly interacts with `libtest`'s output capturing,
and can be configured via `RUST_LOG` when running tests.

this is a tremendously helpful aid in debugging issues during feature
development, and is not specific to the `penumbra-app` crate.

this commit hoists that test utility into a standalone library in
`crates/test/`, so that other components like the dex can also create a
guard to capture traces in tests.

now we can share this, by doing this:

```rust
 #[tokio::test]
 async fn example_test_case() -> anyhow::Result<()> {
     // Install a test logger...
     let guard = set_tracing_subscriber();

     // Test logic here...

     drop(guard);
     Ok(())
 }
```

#### checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > only changes test code