fix(asset): 🔊 `Metadata: Debug` includes all fields
fixes #4190.

> The `Debug` impl for `asset::Metadata` throws away all data except for
the base denom. This is extremely inconvenient for debugging.

- https://github.com/penumbra-zone/penumbra/issues/4190

---

this commit updates `Metadata`'s impl of `std::fmt::Debug` so that it
includes all of the fields in the `Inner` structure.

note for review: #4190 didn't explicitly mention which fields to
include, so i have included all of them as a starting point. if we
should find a middle ground, i am open and interested in hearing about
that!