fix(asset): 🔊 `Metadata: Debug` includes all fields (#4207)
fixes #4190.

> The `Debug` impl for `asset::Metadata` throws away all data except for
the base denom. This is extremely inconvenient for debugging.

this commit updates `Metadata`'s impl of `std::fmt::Debug` so that it
includes all of the fields in the `Inner` structure.

**note for review:** #4190 didn't explicitly mention which fields to
include, so i have included all of them as a starting point. if we
should find a middle ground, i am open and interested in hearing about
that!

#### issue ticket number and link

#4190

#### checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > debug representations should not affect consensus.