ci: fail on broken links in guide
We weren't failing on the broken links in CI, but now we are. This
solves most of our problems: broken `include` statements still don't
throw an error, but they emit a warning. There's an upstream PR to
mdbook to update this behavior, which I've +1'd.

Also patches a few broken links here and there, related to reorg in GH #4215.