auction: `ActionDutchAuction*` validation and component impl (#4220)
## Describe your changes

This PR adds validation logic for:
- `ActionDutchAuctionSchedule`
- `ActionDutchAuctionEnd`
- `ActionDutchAuctionWithdraw`

and a component implementation, on a model similar-but-simpler to the
DEX:

- a `DutchAuctionManager` extension trait defines a crate-level API to
effect a validated action
- an `Inner` internal implementation manages writing to the state
- internal data can be accessed via a general `AuctionStoreRead` trait
which manage tentatively deserializing raw auction state (wkt `Any`)
into a useful domain type.

## Issue ticket number and link

#4211 

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

> Those actions are not integrated in our transaction system, our
application impl.