dex: correct swap claim bsod check
This is a cherry-pick of PR #4239, by:
Author:    Lucas Meier <lucas@cronokirby.com>
Date:      Thu Apr 18 14:55:50 2024 -0700

This PR:
1. fix a bug in the swap claim circuit
2. includes a migration routine for testnet 71

Swap claim proofs take a `BatchSwapOutputData` as part of their
public inputs. This data is used to compute a user's pro-rated share
of a batch swap at a given block.

For that reason, it is important the output data used in the proof
be correct. A swap claim must be bound to the specific BSOD that
was produced by that user's swap.

The swap claim circuit's validation only checked that the supplied
BSOD was produced at the same relative block height as the user's
swap. It did not check that the epochs were the same.  As a result,
it was possible to generate swap claim proofs using a BSOD produced
during a different epoch, on potentially much more advantageous
terms.