feat(governance): 🧊 test that community pool spend proposals can be disabled (#4222)
fixes #4079. related to, but does not finish #2105.

this introduces test coverage to show that community pool spend
proposals can be disabled via a chain parameter.

### related changes

some changes made, related to writing tests for the governance
component:

* 6b3be5892 refactor(governance): 💶 `Value` accessors for proposal
actions
* de16dace5 feat(tx): 🌮 `ActionPlan: From` for community pool actions
* 4e6b874ed feat(amount): ➖ `Amount: std::ops::SubAssign`

### tests

this addresses part of #2105. we should add tests that exercise the
other
tallying logic before that issue is complete. tests are added in the
following
commits:

* fedfa8ded tests(app): 💭 test a community pool spend
* 201fe4873 tests(app): 🏊 test a community pool deposit
* 8543869d4 tests(app): 🥶 test that community pool can be frozen