docs: clarify chain upgrade procedure
Follow-up to #4097. We performed a chain upgrade on Penumbra
for v0.71.0. Based on feedback from the community validators, we've
refined the procedure somewhat. Substantive changes to the docs are as
follows:

  * backup the entire pd & cometbft data dirs, as a precaution
  * upgrade to new pd early, prior to running export, for simplicity's
    sake
  * only move the exported rocksdb directory, which allows preservation
    of other dirs in pd state, e.g. acme tls info
  * do not reset priv validator state
  * amend custom find command to preserve `tx_index.db`

Not yet included in these docs are specifics about the upcoming release;
we'll add those soon.

Finishes and therefore closes #1812.