feat(app): 🎣 add missing `Dex` component hook
see #3799, and #4246.

while wiring up a pair of missing component hooks for the `Sct`
component, i found another (noöp) hook that we do not include in the
`App`'s control flow.

this isn't currently a bug, since `<Dex as Component>::begin_block()`
is a noöps but could easily lead to a surprising outcome in the future, if
more logic was introduced to that trait method.