feat(app): 🎋 add missing `Sct` component hooks (#4246)
fixes #3739.

the `App` only ever invokes the Sct component's `init_chain` and
`begin_block` hooks.

this isn't currently a bug, since its `end_block` and `end_epoch` hooks
are noöps, but could easily lead to a surprising outcome in the future,
if more logic was introduced to those trait methods.

#### checklist before requesting a review

- [x] if this code contains consensus-breaking changes, i have added the
"consensus-breaking" label. otherwise, i declare my belief that there
are not consensus-breaking changes, for the following reason:

> while this **does** change the control flow of consensus-critical
logic, this does **not**
change the execution of incoming data. the hooks introduced here are
noöps.