app: 🧙 `ValidatorDataReadExt::validator_definitions()`
these methods are tremendously useful for tests, because they save us
the trouble of repeatedly hassling with a `Stream`. to discourage future
use of them in production code however, we can define these in a
test-specific extension.

this commit introduces `ValidatorDataReadExt`. `validator_identity_keys`
is also moved, as it is subject to the same performance considerations.