docs: re-add priv val state munging for upgrades
We recently removed mention of zeroing-out `priv_validator_state.json`
for CometBFT as part of the chain upgrade docs. During migration testing
on a private fork today, though, several of us observed that without
this change, the network would not resume. More research required to
understand the root cause here.

Refs #4245, #4239.