docs: chain upgrade for testnet 72
Bumps the application version to v0.72.0, which is required
because the changes in #4293 were merged into a release branch,
and thus didn't land on main naturally. We'll backport that patch
shortly.

More relevantly, the chain-upgrade docs now reference the "current"
migration, which is 71 -> 72, with its own genesis time.

Finally, a small change is made to the recommended dir names, using
unique-to-this-migration dirnames for the export, to avoid confusion
for operators who have previously migrated and not cleaned up the state.