view(proto): remove reserves from auction withdrawal `TransactionPlannerRequest` (#4270)
## Describe your changes

~~Accidentally defined the `TransactionPlannerRequest` auction withdraw
submessage to specify the user reserves in `Amount`. But we need those
to compute a balance commitment, and this requires the asset id.~~

@hdevalence noticed that we can defer populating this data to the view
service (which is where we would have gotten it from in the first
place!)

## Checklist before requesting a review

- [x] If this code contains consensus-breaking changes, I have added the
"consensus-breaking" label. Otherwise, I declare my belief that there
are not consensus-breaking changes, for the following reason:

  > This is proto-breaking.

---------

Co-authored-by: Lucas Meier <lucas@cronokirby.com>