chore: bump chain id for testnet 73
We mistakenly broke testnet 72 on chain upgrade, by leaving a forked
network running that caused tombstoning. To resolve, we're going to
reset chain state and use a new chain id. This commit ensures that the
chain id is bumped `penumbra-testnet-deimos-6` ->
`penumbra-testnet-deimos-7`. Not strictly necessary, but for posterity,
copied over the same configs for the intervening 71 & 72 testnets.