keys: 👯 `Address` is not `Copy`
the core of this commit is here:

```diff
 /// A valid payment address.
-#[derive(Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(try_from = "pb::Address", into = "pb::Address")]
 pub struct Address {
```

out of that, a few other types also lose their `Copy` property:
* AddressView
* FundingStream
* Recipient

this commit looks noisy, but it is not changing any behavior. these are
the same copies as previously made, but are now done explicitly. see:
https://doc.rust-lang.org/stable/std/marker/trait.Copy.html#whats-the-difference-between-copy-and-clone

many of the changes here affect test code.